--- a/.gitattributes
+++ b/.gitattributes
@@ -3,5 +3,8 @@
 # Mark the database schema as having been generated.
 db/schema.rb linguist-generated
 
+# Mark the package-lock file as generated
+package-lock.json linguist-generated
+
 # Mark any vendored files as having been vendored.
 vendor/* linguist-vendored