--- a/.github/workflows/rubyonrails.yml
+++ b/.github/workflows/rubyonrails.yml
@@ -36,7 +36,7 @@ jobs:
         with:
           bundler-cache: true
       - name: Install frontend dependencies
-        run: yarn
+        run: npm ci
       - name: Run Frontend build
         run: bin/vite build
       # Add or replace database setup steps here
@@ -58,7 +58,7 @@ jobs:
         with:
           bundler-cache: true
       - name: Install frontend dependencies
-        run: yarn
+        run: npm ci
       # Add or replace any other lints here
       - name: Security audit dependencies
         run: bundler exec bundle audit check
@@ -67,4 +67,4 @@ jobs:
       - name: Lint Ruby files
         run: bundler exec rubocop --parallel
       - name: Lint Javascript and Sass files
-        run: yarn lint
+        run: npm run lint