--- a/.gitignore
+++ b/.gitignore
@@ -47,3 +47,5 @@ config/environments/*.local.yml
 
 # Vim swap files
 *.swp
+
+/public/assets