--- a/Procfile.dev
+++ b/Procfile.dev
@@ -1,2 +1,2 @@
 web: bin/rails server -p 3000
-frontend: yarn dev
+vite: bin/vite dev