--- a/app/assets/config/manifest.js
+++ b/app/assets/config/manifest.js
@@ -1,3 +1,0 @@
-//= link_tree ../images
-//= link_tree ../builds
-//= link_tree ../../../node_modules/govuk-frontend/govuk/assets