--- a/app/components/summary_card_component/_index.scss
+++ b/app/components/summary_card_component/_index.scss
@@ -12,7 +12,7 @@
     @include govuk-media-query($from: desktop) {
       display: flex;
       justify-content: space-between;
-      align-items: start;
+      align-items: flex-start;
     }
 
     @include govuk-media-query($media-type: print) {