--- a/app/components/task_list_component/_index.scss
+++ b/app/components/task_list_component/_index.scss
@@ -42,6 +42,7 @@
 
 .app-task-list__item {
   border-bottom: 1px solid $govuk-border-colour;
+  // stylelint-disable-next-line declaration-no-important
   margin-bottom: 0 !important;
   padding-top: govuk-spacing(2);
   padding-bottom: govuk-spacing(2);
@@ -58,7 +59,6 @@
     float: left;
   }
 }
-
 
 // The `app-task-list__task-completed` class was previously used on the task
 // list for the completed tag (changed in 86c90ec) – it's still included here to