--- a/app/frontend/entrypoints/application.js
+++ b/app/frontend/entrypoints/application.js
@@ -1,6 +1,5 @@
-// Entry point for the build script in your package.json
 import { initAll } from 'govuk-frontend'
-import copyToClipboard from './modules/copy-to-clipboard'
+import copyToClipboard from '../javascript/copy-to-clipboard'
 
 document
   .querySelectorAll('[data-module="copy-to-clipboard"]')