--- a/app/frontend/entrypoints/application.scss
+++ b/app/frontend/entrypoints/application.scss
@@ -1,8 +1,8 @@
-$govuk-images-path: "./images/";
-$govuk-fonts-path: "./fonts/";
+$govuk-images-path: "@govuk/assets/images/";
+$govuk-fonts-path: "@govuk/assets/fonts/";
 
 @import "govuk/all";
-@import "modules/app_select_options";
+@import "../styles/app_select_options";
 @import "../../components/task_list_component/";
 @import "../../components/page_list_component/";
 @import "../../components/summary_card_component/";