--- a/bin/setup
+++ b/bin/setup
@@ -29,7 +29,7 @@ FileUtils.chdir APP_ROOT do
   system! "bin/rails log:clear tmp:clear"
 
   puts "\n== Installing frontend dependencies =="
-  system! "yarn"
+  system! "npm install"
 
   puts "\n== Building frontend assets =="
   system! "bin/vite build"