--- a/config/vite.json
+++ b/config/vite.json
@@ -1,0 +1,18 @@
+{
+  "all": {
+    "sourceCodeDir": "app/frontend",
+    "entrypointsDir": "entrypoints",
+    "watchAdditionalPaths": [],
+    "additionalEntrypoints": ["node_modules/govuk-frontend/govuk/assets/**/*"],
+    "publicOutputDir": "assets",
+    "assetsDir": "",
+    "autoBuild": false
+  },
+  "development": {
+    "port": 3036
+  },
+  "test": {
+    "port": 3037,
+    "autoBuild": true
+  }
+}