--- a/package.json
+++ b/package.json
@@ -12,40 +12,23 @@
     "Safari >= 12"
   ],
   "scripts": {
-    "build": "gulp javascript",
-    "build:css": "gulp css",
-    "dev": "gulp dev",
     "test": "jest",
-    "lint": "npm run lint:js && npm run lint:css",
-    "lint:js": "standard | snazzy",
-    "lint:css": "stylelint 'app/assets/**/*.scss'",
-    "format": "standard --fix & stylelint 'app/assets/**/*.scss' --fix",
-    "ci": "bin/setup && bin/rails db:seed && API_KEY=test bin/rails server"
+    "lint": "standard | snazzy & stylelint 'app/frontend/**/*.scss'",
+    "format": "standard --fix & stylelint 'app/frontend/**/*.scss' --fix",
+    "ci": "bin/setup && bin/rails server"
   },
   "devDependencies": {
-    "@babel/cli": "^7.17.6",
-    "@babel/core": "^7.17.8",
-    "@babel/preset-env": "^7.16.11",
-    "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-commonjs": "^21.0.3",
-    "@rollup/plugin-node-resolve": "^13.1.3",
     "autoprefixer": "^10.4.4",
-    "core-js": "^3.21.1",
     "cssnano": "^5.1.7",
-    "exec-sh": "^0.4.0",
-    "gulp": "^4.0.2",
-    "gulp-postcss": "^9.0.1",
-    "gulp-sass": "^5.1.0",
     "jest": "^27.5.1",
     "postcss": "^8.4.12",
-    "rollup": "^2.70.1",
-    "rollup-plugin-command": "^1.1.3",
-    "rollup-plugin-terser": "^7.0.2",
     "sass": "^1.49.11",
     "snazzy": "^9.0.0",
     "standard": "^16.0.4",
     "stylelint": "^14.6.1",
-    "stylelint-config-gds": "^0.2.0"
+    "stylelint-config-gds": "^0.2.0",
+    "vite": "^4.2.0",
+    "vite-plugin-ruby": "^3.2.0"
   },
   "dependencies": {
     "govuk-frontend": "~4.5.0"