--- a/package.json
+++ b/package.json
@@ -15,7 +15,8 @@
     "test": "jest",
     "lint": "standard | snazzy & stylelint 'app/**/*.scss'",
     "format": "standard --fix & stylelint 'app/**/*.scss' --fix",
-    "ci": "bin/setup && bin/rails server"
+    "ci": "bin/setup && bin/rails server",
+    "dev": "vite dev"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.4",