--- a/spec/helpers/application_helper_spec.rb
+++ b/spec/helpers/application_helper_spec.rb
@@ -129,4 +129,10 @@ RSpec.describe ApplicationHelper, type: :helper do
       end
     end
   end
+
+  describe "#govuk_assets_path" do
+    it "returns the full node_modules asset path" do
+      expect(helper.govuk_assets_path).to eq "/node_modules/govuk-frontend/govuk/assets"
+    end
+  end
 end