… Gemfile.lock
… app/frontend/javascript/copy-to-clipboard/index.js
… app/frontend/styles/_app_select_options.scss
… bin/vite
… config/initializers/assets.rb
… gulpfile.js
… yarn.lock