@danworth 2023-03-21
This all seems well considered, I don't know enough to offer an expert review though. Has this been deployed to PaaS dev and or AWS dev (deploying to AWS dev would ensure the docker image is built correctly and runs as expected when the environment is set to 'production' ). If not and you think it's necessary I'm happy to run through how to use the AWS pipelines to deploy dev branches to AWS dev environment, just let me know.