@DavidBiddle 2023-03-21
@danworth It's been up to PaaS dev without any issues, I'm up for checking it on AWS dev as well if that's useful. I _think_ I can do that myself now but can shout if I get stuck