@lfdebrux 2023-03-22
Note that one might need to run `bin/setup` to get Vite to work (I did)